requirements.txt
discord.py[voice]==1.2.3
youtube_dl==2019.7.2
requests==2.22.0
praw==6.3.1
pygithub==1.43.8
requests
